This commit solves some little bugs in the iPing application.
1) The client sends to the server a name containing (among the other things) the size of the ContentObject the server has to reply to the client.
The problem is that the Server is expecting to find this size always at the same position in the name inside the interest; if the client sends an interest with a name that matches the FIB but is not exactly the name the server is expecting (e.g. ccnx:/ping/a instead if ccnx:/ping), the server reads the wrong segment (not the one with the size) and sends out a content object with an undefined size.
2) The maximum payload size of a content object is 64000. This causes fragmented packets to not be processed by the forwarder itself if they are coming from a remote connection. A quick fix consists in setting the max payload size of a content object to 1400.
So the max size of a content object should be 1400 and not 64000.

Change-Id: Ib555b7beb3cfe9172904bd63b8e3a76fe3f71a50
Signed-off-by: Mauro Sardara <msardara+fdio@cisco.com>