[CICN-27] Fix invalid pointer dealloc
There is a deallocation of an invalid pointer in the destructor
of an in-memory verifier object. This function is called by the
destructor of consumer socket which fails because of that. The
bug is visible only on Ubuntu 16.04.

The bug was introduced in CICN-26 as an artifact of attempts to
fix memory leaks in Android. There is no good reason to keep the
changes causing the problem and the concerned lines can be safely
removed.

Signed-off-by: Olivier Roques <oroques+fdio@cisco.com>
Change-Id: Ibcde7dd7510d76f5aa91e5967a4261ec7eb49108